requirements.txt
# latest @ Feb 2025
# Note updating to Python 3.12.7 has caused multiple wheel failures. 

# APP
python-dotenv==1.0.1
azure-storage-blob==12.24.1



# =====================================


# WORKING
#dash==1.18.1 #Dash 2.0 problematic and unneeded. Still problematic on firefox on linux.
#flask==2.0.3 # Note Flask 3 is out now
#Werkzeug==2.0.0 # needed due to weird flask error
#kaleido==0.2.1 #pdf export
#dash-deck==0.0.1
#pydeck==0.4.1 #Had to lock this as updates are breaking. Might be cool new features though.
#gunicorn==20.0.4
#dash-extensions==0.0.58 #Download features
#dash-bootstrap-components==0.13.1 #this is as high as we can go before dash 2.0 breaking changes come into effect
#xlrd==1.2.0
#xlsxwriter==1.4.4

# UPDATED
#pandas==2.2.3
#pillow==11.0.0
#matplotlib==3.9.2
#pyarrow==18.0.0
#azure-storage-blob==12.23.1


# LAKEHOUSE
#openpyxl==3.1.5 #Needed for opening xlsx files using pandas. 

# SPYDER
#spyder-kernels==3.0.1

